+++ tests/Feature/ConnectionControllerTest.php
@@ -10,7 +10,7 @@ class ConnectionControllerTest extends LaravelTestCase
     /** @test **/
     public function can_get_available_connections(): void
     {
-        $response = $this->getJson('report/connection');
+        $response = $this->getJson(route('report.connection.list'));
 
         $response->assertStatus(200);
     }